Handle long strings in build/buildsets tables
A few user-supplied strings can have an adverse effect on the build
and buildset tables.  Normally these fields wrap if the strings are
too long, however there are some circumstances where they will not
wrap.  Especially if more than one of these applies, it can cause
entirel columns to be pushed out of the right side of the viewport
and be inaccessible.  Browsers will wrap words on spaces and hyphens,
but extra work is needed to allow them to wrap in other cases.
Some example strings that cause this problem are:

* job-nameswithreallylongwords-in-them
* branches/with_underscores_instead_of_dashes
* changes,withcommitshasaspatchnumbers

To address all of the columns except changes, we add the patternfly
breakWord cell transform.  This will allow the text to break anywhere
necessary (but will still prefer spaces and hyphens).

For the change column, we set the width of the column to 10% and
truncate the text within.  In practice, this appears to allow for about
10 digits or so.  That's enough to fully identify a typical gerrit
change and patchset, as well as effectively uniquely identify a
github PR and commit (at least, unique enough within the table display).
The rest will be truncated with elipses (...).

Change-Id: I067528af64dd106353dc6be877dec938a0d2e8e0