Fix config error detection on same project branch
The following commits (changes):

050c640b0bab32c787756147c8a470e1fbb0206f (I885a2d3ab6d60402425b879e639c1e597717622a)
0fc385ebed38f8d49d275bf39403b723a462e649 (I8a72e57e068b073e37274cdabeacffc9daee2e94)

established the idea that we should not report configuration errors
that already exist unless they are present on the project-branch of
the change under test.  That is to say, we should report:

* Errors in other projects or branches of this project which did not
  exist before this change.  I.e., removing a job from this repo which
  breaks another repo should be reported.
* Errors in this project-branch.

But both of these only under the condition that the configuration
is being changed in this patch.

We could almost get away with not reporting the second condition,
and in fact, because of the bug which is fixed in this commit, we
are not currently reporting the second condition.  But we should
because we can't tell with perfect fidelity whether an error in this
project-branch which existed before this change and also exists in
this change is really the same error.  Or if there are two errors,
they may fix one but not realize that the second remains.

This patch restores the original intent of the code and corrects a
bug where we compared a project object to a project name (thereby
rendering the condition always false).

This will require users to, if they change Zuul's configuration,
ensure that all errors on the project-branch of that change are
resolved in order for Zuul to validate the change.  This does not
affect changes which do not change the Zuul config.

Change-Id: I50c37a621be6c2f2929284b9af994b36eca2ba8f