Update patch set 1
Patch Set 1:

So is it timing out on reporting?

We added retries to the gitlab driver, but there was a bug in that change which we only just merged a fix for yesterday.  If you aren't running that fix, then it's possible that retries aren't actually happening.  Let's make sure that's working first before we change this.

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>