Handle no file commits in the Gerrit driver
It is possible to push no file commits to Gerrit. We had previous code
looking for this case by checking for the special '/COMMIT_MSG' file
entry as the only entry in change.files. Unfortunately, the gerrit http
connection queries were returning a dict of files in the form of
{'name': {attrs}' and this broke the special checking for the single
commit message entry.

We fix this by coercing the type of the returned files data from a dict
to a list of file names.

Change-Id: Ie475911b1c9668bd831c4d31213a2fd3a8c049a7