Quickstart wait for certs in quickstart docker-compose.yaml
Zookeeper, Zuul, and Nodepool services need the zk certs to be present
before starting. Without this they cannot communicate with each other
with TLS which is required. Docker-compose doesn't do a strict startup
ordering waiting for each service to be ready. It just ensures processes
begin in the right order.

Fix this with a new wait script and override container commands to run
the wait script as necessary to ensure certs are present before we
begin.

Change-Id: I8179159ae7d6a15155066549dfe245607646d433