Add TCP keepalive to gitlab
The gitlab driver uses requests to interact with gitlab's API.  By
default, requests uses a connection pool to cache and re-use HTTP
connections. If the gitlab server is across a network segment with,
say, a poor NAT implementation which silently drops idle connections
from its map, then Zuul may not notice until it attempts to make a
request on one of these connections and receives a read timeout.

To avoid this, use TCP keepalives on these HTTP requests by default.

This is configured to send a keepalive packet by default every 60
seconds.  This should avoid sending too much extra traffic, while
improving the reliability for all users.  The default of 60 seconds
was chosen to match the SSH keepalive value we set for gerrit
connections (for much the same reason).

If this works well, we can use this approach with other requests-based
drivers (and at this point, every driver uses requests to at least
some degree).

Change-Id: I86ce064c6ec9325e2ae9db516778058050696ef6