Update patch set 6
Patch Set 6: Code-Review-1

(1 comment)

-1 for the reno.

I'm sort of at 0 on the change itself.  While I agree this is some work we can do for users automatically and it may help in some cases, it strikes me that just as often as changing a job's playbook is changing a role.  But we're not planning on parsing the playbooks themselves to identify what roles they use, so users still need to list roles in file matchers.  Automating the playbook addition might lead people to believe that any change to the ansible code should run the job.

I will note that the situation is a little different from the job config change.  In that case, the feature is urgently needed since in many repos, all the jobs are defined in a single file, and listing "zuul.yaml" in a files matcher in a repo would mean any change might run all jobs which is rarely what is desired.

Patch-set: 6
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1