Resolve cache keys when gathering related changes
When we clean up the connection cache, we exclude changes which
are in pipelines and any changes related to those.  Gathering the
related changes is a partially recursive function.  When we switched
to using cache key references for keeping track of dependencies
between changes (rather than direct in-memory object references),
we did not update the method which recursively walks the
reverse-dependency path to collect that set.  That caused an error
since it was expecting to call a method on a change object but instead
was given a change key.

To correct this, we resolve the change key to a change object before
calling that method.

We were also missing some exception hnadlers, so this exception was
not being recorded in the scheduler log.  That is corrected and some
additional diagnostic debug log messages are added.

Change-Id: I6ac2e2b63c71d609d9f023217f2a435d18f1159d