Add RepoState object
This makes the repo_state buildset attribute a pair of RepoState
objects so they can be stored in ZK as sharded data (since the repo
state can be quite large).

It's done as a pair because we perform up to 2 merge operations to
create the repo state.  Storing the results individually means that
we can recover in the case where we crash while writing them out.

Change-Id: I586ba591fd489279514ef2c86830391a8b8ac32d