Fix race in test_node_request_canceled
Even though our nodepool is paused, it still sees events and so adds
them to the provisioned_requests list.  This test only passed before
in the cases where it ran very quickly and beat the arrival of the
event.

To make this safer, allow the event to arrive and assert that even
after the request is deleted, we still only saw one request and
it remains in 'requested' state (ie, it was never fulfilled after
we canceled it).

Change-Id: I92996842e1205c7cde880c20d3350c05b2ca1edf