wip: Use a deterministic build set UUID
The build set defines the current position of an item in the queue. If
the item is moved the build set is no longer valid.

We use the UUID of the build set ahead for and the item UUID to generate
a deterministic UUID for a build set. This allows us to detect if a
build set is still valid.

Open issue:
* cancel jobs / reset build set in case the position of the item change
  during re-enqueue
* re-enqueueing items is currently not considering the nnfi, but will
  just sequentially enqueue items. This means that we might cancel jobs
  that don't need to be canceled since the items ahead/behind are not
  correct for the current sequential state of the queue but would be
  correct if we considered the nnfi.

Change-Id: I742af0e892d640d59b6ace637158f13183ded2d9