gerrit: handle POST 400 errors and don't retry
A 400 response from Gerrit indicates that we sent something bad, or
that something failed validation, etc.  It is not going to disappear
with retries.  Thus this catches 400 separately on POST requests and
stops any retry loops.

The response also returns some text telling you what is wrong; log
that in the exception (it is logged at the zuul.GerritConnection.io
logger, but that is usually set high to avoid too much log traffic).

Change-Id: I6a470bf338cf8b9944760c89d5742c79318c7cb8