gerrit: trim messages to "human length"
The "message" field of a gerrit response (i.e. the bit that has "Zuul
encountered a syntax error while parsing its configuration ...") has a
limit in Gerrit of "change.commentSizeLimit" which is by default 16KiB
[1].

If you have a syntax error in a very long section, say the middle of
the project: section of OpenDev system-config, you can actually
overflow this length.  Gerrit rejects the update and as an end-user
you don't get any clue why.

This truncates the message in the gerrit driver if it's going to
overflow.

[1] Note change.robotCommentSizeLimit is larger, but that's not how
syntax messages are reported.

Change-Id: I2f14e734ef5f9f203b14556c7d2c8ed1ad052319