Store and resolve queue item's ahead/behind refs
The item(s) ahead/behind references of a queue item are now stored in
Zookeeper. The references will be resolved after we've refreshed all
items in the queue during deserializing the `ChangeQueue`.

The reason for resolving the references in the change queue instead of
the item itself is that we might not have created the referenced items
at the point when we deserialize a single queue item.

Change-Id: I15aaa3026b606cd99ecc3ac013062ad35b2fbbb8