Move job_graph attribute to BuildSet
Currently the job_graph is an attribute of a QueueItem.  However
it's really a characteristic of a BuildSet.  It has a 1:1
correspondence with a BuildSet.

This doesn't make a big difference in the scheme of things, but
I think it does help keep our mental models in sync with the data
models.  The only practical outcome of this is that we don't need
to explicitly zero-out the job_graph attribute on QueueItem when
we create a new BuildSet.  Otherwise, this change is effectively
a no-op.

But since we're about to serialize the jobgraph data, I think it
makes sense to attach it to the object it's closest to.

Every remaining attribute of the QueueItem is independent of the
BuildSet.

Change-Id: Iaa8dd37d73471223459f3179bf38c87f95a6fd38