Serialize JobGraph objects to ZK
These are serialized as attributes of the BuildSet.  Essentially,
the JobGraph is a logic encapsulation class with a small amount of
data.  These extra data are stored on the BuildSet object, and
the JobGraph is reconstructed from them as necessary.  This saves
extra ZK network traffic for a small amount of data that has the
same lifecycle as the BuildSet.

It might be structurally simpler to just move all of this into the
BuildSet directly, but the logic is fairly complex and I think
benefits from staying in its own class.

The ProjectMetadata object needs to be serializable to complete
this.

Change-Id: Id451d2b24556f62927120ccdd2657772f12c787b