Create Abstract and FrozenJob classes
In preparation for storing the "Job" in ZooKeeper, this creates
an AbstractJob and FrozenJob class.

The FrozenJob will, in a subsequent change, be serialized into ZK.
This is the rendered job definition that the build will run, plus
it has some state variables (like whether it's waiting on nodes
or dependent jobs, etc).

The existing Job class will continue to be the in-memory job
definition that is part of the layout.

Change-Id: I0e2a017fc7abd9e2346035816223c585eac9f422