Remove setBase from job freeze API
As part of the work to freeze jobs into ZooKeeper, it was noted that
the freeze job API called setBase on a frozen job.  This is not what
happens during actual job freezing and therefore the API produces
output that differs from what would actually be run.  The setBase
method is already called at the start of the job freezing process and
sets up frozen playbooks and the initial inheritance path for the
first variant that is applied.  Subsequent variants add their
playbooks using applyVariant.  It is not clear why setBase was being
called a second time after all variants are applied.  By removing this
call (which should not be necessary on a frozen job) the only change
in the test output is to remove a non-sensical item from the end of
the inheritance path (job project-test1 on common-config/zuul.yaml
line 53 -- that line is actually the 'base' job; it correctly appears
as the first entry in the list).

To resolve this, the apparently unnecessary setBase call is removed
from the freeze job API.

Change-Id: Ibb2bb8d17bf1a665467affe583f2b7780fb54322