Remove toDict from FrozenJob
The only use of FrozenJob.toDict is to compare a previously frozen
job to a not-yet frozen job to see if the configuration has changed;
this is why toDict is on the AbstractJob -- both kinds of jobs need
to return the same info in toDict for this comparison.

If, instead, we calculate a hash on the config Job and then also
store that hash on the frozen job, we can compare the two without
needing the full toDict method on the frozen job.  This simplifies
the frozen job considerably (as it no longer needs to have all of
its children in object form just for the purposes of this comparison).

This leaves us with no commonality in AbstractJob, so it is now
removed.

Change-Id: Idf4c97baf041a40e90acfcc38f39b4d7836d3b4f