UI: Fix build time calculation for empty buildsets
This fixes a bug on the buildset result page which prevents the page
from loading successfully due to the following exception:

  Uncaught TypeError: Cannot read properties of undefined (reading
  'reduce')

This exception is raised when we try to find the first and last build
of an empty buildset (that doesn't contain any builds) because the
builds attribute is undefined.

  > 43 | const firstStartBuild = buildset.builds.reduce(
    44 | ^ (prev, cur) => (prev.start_time < cur.start_time ? prev : cur)
    45 | )

To fix this change, we move the time calculation and the related
rendering parts into a function and only call it if the buildset
contains builds.

Change-Id: I1694b8dd26d27bd0d4d5f65d5e057af860729ad6