UI: Ignore empty timestamps in build time calculation on buildset page
Currently, if a buildset contains at least one build without sufficient
time information (e.g. a SKIPPED build), this build will be shown as
"Starting build" with an "Invalid date". Additionally, the calculated
duration of such a buildset will always be shown as 0.

To fix this behaviour, we simply ignore missing timestamps when
trying to find the first and last build in the buildset.

Change-Id: I9a4479c913743ebf1fc8f0133f544362dcf61aaf