Add addtional checks to key deletion testing
It occured to me that we should also test that the removal of one
project's from an org in zookeeper should not remove other projects or
the org itself. We only want to remove a project whose keys are removed
and an org when all projects are removed from it.

Change-Id: I5bb3192785fe8a863b82f7d13494bd330541f0a1