Update patch set 5
Patch Set 5:

James E. Blair wrote:

> Oh, and there's no link to the tenants page from the components
> page; we should probably add that.

I guess the "zuul" icon navigates there, but I'm so used to clicking the "Tenant" link in the header to go back there that I didn't think to do that.  I'm open to the idea that I should just learn to click "Zuul" instead, but I think my initial reaction is legitimate, and maybe we should put "Tenants" in the same spot as "Tenant: {name}" on both the tenant and component pages?

Patch-set: 5