Use a ThreadPoolExecutor for kazoo callbacks
The kazoo client can spawn a number of background threads to perform
async tasks without blocking foreground operation.  In particular,
this allows its main thread to keep up with ZK traffic which may
trigger callbacks, and then if those callbacks need to perform
operations which take time (or cause ZK traffic themselves) they
will spawn threads for these operations.

Watches set session callbacks which refresh their data.

If a ZK session is reset, then kazoo will call all of these callbacks,
each of which may spawn a thread to refresh its data.  This can easily
lead to spawning too many threads and/or running out of memory.

To alleviate this, use a threadpool executor in the kazoo client when
spawning background operations.  We allow 10 workers for this.  The
watches (which we vendor) are updating so that their throwaway operations
use the thread pool executor, but the long-running background threads
are top-level threads as usual.

Change-Id: Ie90f904342f6261859a875e75b1ea14e2238f895