Create bundle items during queue deserialization
Since `Bundle` items represent a circular dependency we need to be
careful not to run into infinite recursion when re-constructing the
queue.

In addition all items in a bundle need to share the same `Bundle` object
instance. A newly added bundle UUID makes it easy to indentify items
that should share the same bundle.

To guarantee this we will reconstruct the bundle when the change queue
is deserialized. This is similar to how we set the items ahead/behind
references of a queue item.

Change-Id: I32b1111fcb0902bbee70b6fee3415f7dda56c351