Ensure same layout UUID across schedulers
Until now the layout UUID was randomly generated on each scheduler.
However, since we now use the layout UUID to detect if we need to reset
the pipeline state, the UUID must be the same on all schedulers.

Having the same layout UUID on all schedulers might also make debugging
easier in some cases.

For this we will store the layout UUID as the UUID of the layout
state in Zookeeper.

Change-Id: I400b96b54b3dc9e30fb58e56e1e25a79427b5e70