Only reset the pipeline state if needed
Always resetting the pipeline state in the manager's post-config method
doesn't work, as it will be executed on each scheduler.

In order to detect if we need to reset the pipeline, we will save the
current layout UUID in the pipeline state and use that to detected an
outdated state.

Change-Id: Ifb87417b792025b4d9723b5dda5f508c3b9c7e4a