Use a metaclass to deserialize event objects
We need a mapping of class names to class objects for all of the
different event types we might serialize to ZK, and it's tedious
to maintain that by hand.  Instead, add a metaclass which records
each class name/definition in a dictionary as each subclass of
AbstractEvent is defined.

(This will also add AbstractEvent to the mapping, but that's
harmless.)

Change-Id: Ic0701be49820cfd1dee15ea48f50935fa217ea36