Fix race is test_zookeeper_disconnect
If it takes too long for the scheduler to win the nodepool request
watcher election after the simulated disconnect, then the fake
nodepool may fulfill the request and the scheduler will notice that
but not immediately add the NodeProvisionedEvent to the queue.

If waitUntilSettled checkes during that window, it will erroneously
conclude things are settled.  To avoid this, wait until the scheduler
has won the election before proceeding.

(An alternative would be to have waitUntilSettled check this, but
this may only apply to a simulated zk restart, so that may be too
heavyweight of a solution.  But if we see other similar problems in
the future, we may want to consider that.)

Change-Id: I887a7a0125d659192ea76d84596987b28a2b9f4d