Add a pipeline change list object to ZK
This is an optimization so that the scheduler doesn't need to refresh
the entire state of every pipeline in a tenant when processing the
event queues.

We normally decide whether to forward an event to a pipeline based on
the pipeline triggers.  But we also check whether the event is for
a change which is already in the pipeline, because the manager may
need to take action based on that event.  In order to do this,
the scheduler needs to know all the changes in every pipeline (since
it evaluates this for each pipeline).

We don't want to have the scheduler lock every pipeline and refresh;
we're trying to shard processing by only having the scheduler lock
and process one pipeline at a time.

To resolve this, let's add a new object in ZK which is simply a list
of changes in the pipeline.  We will update this list after each
time we process the pipelines, and we will have the scheduler load
this list for all the pipelines when it starts processing the trigger
event queue.  It can do this without a lock.

This will be "best-effort" as there are some race conditions where
we may not forward an event where we should (or vice versa).  We
also only store the first 1024 changes that we see in order to avoid
sharding issues.

Change-Id: I274dcc6f9e6cf7e64153abc66036d5bc24ade25a