wip: Allow refreshing project branches
The connection's project branch cache is normally cleared when a related
event is processes. Since the event processing only happens on a single
scheduler the project branch cache will be outdated on the other
schedulers.

This is a WIP fix attempt that refreshes the branch cache when
(re)loading a tenant. However, this doesn't solve two other cases where
we need the project branch cache to be up-to-date:

* checking if a reconfig is needed when forwarding trigger events based
  on the list of cached branches
* generating trigger events in the timer driver

The usage in the timer driver could be change to use
`tenant.getProjectBranches()` as this will be updated during a
reconfiguration.

Change-Id: I173b4b4c9a24a74373662570073b61081571c2d9