Provide zstat version when updating Node Request in ZooKeeper
With this change, the node request object must be up-to-date before it
can be stored, which avoids outdated data being written to ZooKeeper.

The purpose of this change is a bug in the node request handling which
could reset the node request back to REQUESTED shortly after it was
fulfilled by nodepool.

The problem occurrs when revisedRequest() is called shortly after the
node request was fulfilled by nodepool but the node request cache in the
scheduler is not yet up-to-date. As the node is no longer locked by
nodepool, reviseRequest() will update the data in ZooKeeper with the
outdated information that is stored in the node request object.

As a result the node request in ZooKeeper will be reset to REQUESTED
although it was already set to FULFILLED by nodepool. In this case
neither the scheduler nor nodepool seem to process the node request
anymore. The scheduler thinks it's still in state REQUESTED and
nodepool most probably takes note of the requests it has already
handled.

Change-Id: I713c8c442e2c4022ceaaa5486ee3ab2b359fbfdf