Fix race in change cache removal
The change cache has a childwatch which it uses to do two things:
add data watches for new cache entries so that it can maintain up
to date copies of all items, and also to delete those local copies
when they are removed from the zk cache.

The following race was possible (with thread ids):

1] Child watcher gets notified of a change in ZK
   This call includes the list of children at the time.
2] Cache set call adds an entry to ZK and the local cache.
1] Child watcher gets local cache list, sees an entry in that
   list which is not in ZK and deletes the entry.

Subsequent cache fetches or updates will fail since the item has
been removed from the local cache.

To correct this, we only remove items from the cache if we've
already set a watch on them.  This means that we're using changes
to the watch list to determine deletions from the cache.  This
should be fine for changes which are added to the local cache because
they appear remotely (the watcher is the only way they will be added).
For changes which appear locally it means we won't set a watch on them
until they also show up in the childwatch (which should only be a
few milliseconds later).  That short delay should be fine since we just
inserted the change into the local cache with the most up to date data.

Change-Id: I1d0d9878e10f5be9b6ef96000d535017b3dcdea9