Clean up leaked lock path in request queue
Further to the same problem described in:
I1fc15870c55ca20a586a78c2e9ea3a3bfb58200b

It's still possible for isLocked to see the parent lock path exists,
and then for the lock to be removed atomically in another thread,
and then isLocked tries to lock it, re-creating the lock parent.

This will clean up the lock parent in that case.

Change-Id: Ie9ab6d5306407828d525207df9655a8c564ca062