Use ZK transactions when deleting job request locks
If the lostRequest method is running right as a request was being
deleted, it's possible for the following to occur:

1] lostRequests gets a cached list of requests that are running
2] a request node is removed
2] the request lock node is removed, but the parent node is not yet removed
1] isLocked is called
1] isLocked sees the parent node
2] the child node is removed
1] isLocked creates a lock to check if there are children and therefore causes
   the parent node not to be deleted by thread #2
1] isLocked returns false

This would not be possible if lock deletions were atomic, so attempt
to make them so by using a ZK transaction to delete the lock.  Since
requests are deleted while holding the lock, making the lock deletion
atomic should avoid the race when checking the lock.

Change-Id: I1fc15870c55ca20a586a78c2e9ea3a3bfb58200b