Remove getZKWatches
The client.command() method is limited to returning 8kb responses,
so this check was flawed.  In a busy concurrent test environment,
the zk server may have so many watches that the verbose listing
of them exceeds 8k.

Rather than try to fix this, let's just remove the check.  We may
introduce a leak again by doing this.  If we feel it's worth it,
we could implement our own command method and re-implement this
check.

Change-Id: I815aeee751761c1ecbc17ce180ce73611d98e11f