Remove false-negative in zk test
The method we use to look up watches can return the empty set
even if there are watches on busy systems.  We can fix that, but
it's not a very important check right now, so isn't worth the
effort at the moment.  This change removes the positive assertion,
but leaves the negative one (which when run in isolation should
still produce a useful result if we introduce a watch leak).

Also add a TODO about how to fix this properly.

Change-Id: I815aeee751761c1ecbc17ce180ce73611d98e11f