Update patch set 26
Patch Set 26:

Interim comment: using firefox+keycloak, I had to set the following option:

> OpenID Connect Compatibility Modes
> > Exclude Session State From Authentication Response: On

Otherwise I would end up with a bunch of background round-trips to keycloak.  I have opened this issue with more details to see if that's something that can be fixed in the JS libraries: https://github.com/bjerkio/oidc-react/issues/752

In the mean time, we may want to adjust the import file to include that setting, and also our documentation to mention that.

Patch-set: 26
CC: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>