WIP Use AuthProvider
This makes some changes so that we can use the AuthProvider as
supplied by oidc-react.  Most of the theory is explained in
the comment in ZuulAuthProvider.jsx

The benefit of doing this is that we allow the AuthProvider and
userManager to handle the callback logic, so we don't need a
custom callback page.

The challenge in using these classes as-is is that our authority
endpoints (eg, the IDP itself) may change from one tenant to
the next; these classes aren't set up for that.  So we need to be
careful about how and when we change those authority URLs.

Change-Id: I955f64b8d68909b4547b56798507a6bef6f3af32