Keep consistent types on QueueItem ahead/behind attrs
We serialise item_ahead/items_behind as string paths, and when
we deserialize the queue, we resolve those back into objects.
There is a window during deserialization where we are deserializing
all of the individual items and their attributes are strings, then
we do a pass over all of them and resolve them into objects.

If the status json gearman job runs during this window, it will
throw an exception.  To avoid this, store the strings on an
alternate attribute name, then use that to reconstruct the objects.

Thes restores the previous "best-effort" behavior of producing
a coherent status json.

We will soon be moving this functionality to zuul-web where it can
take further steps to ensure consistency.

Change-Id: Iec92e16e018bdecf21e868cf91110549090cfe31