Fix change cache relevance detection for refs
The getRelatedChanges method is supposed to add the change itself
to the related changes, but we only did that for Change objects,
not any other kind of Ref.  That could cause tags, branch tips, or
other non-change refs to be evicted from the cache if they were
attached to long-running jobs.

Change-Id: I80677cf09915bb8f96c1ac14e309007f7ab7bf69