Use pipeline change list for cache maintenance
In case the connection cache maintenance runs right after a tenant
reconfiguration it might happen that we clean up changes that are still
relevant. This happens when the pipeline wasn't re-enqueued yet and the
queues are still empty.

To fix that we will use the pipeline's change list to collect the
cache keys that are relevant.

Change-Id: Ie79b6110ae26c1d7915aa6c17bed4f613a530794