Implement management events directly in zuul-web
This implements the functionality for the enqueue, dequeue and promote
API endpoints directly in zuul-web, so there is no need to call the
scheduler via RPC.

This does not remove any RPC methods as those are still used by the
zuul command line client.

Those will most probably be removed in a dedicated step later on as
there is now an independent zuul-client[1] package available.

[1]: https://opendev.org/zuul/zuul-client

Change-Id: I0da91b3e80ab9ccb134b9a37f88305fb24533b38