Handle management events directly in zuul-web
With this change, zuul-web puts enqueue, dequeue and promote events
directly in the respective ZooKeeper event queues without taking the
detour via the scheduler.

This does not remove any RPC methods as those are still used by the
zuul command line client. Those will most probably be removed in a
dedicated step later on as there is now an independent zuul-client [1]
package available.

[1]: https://opendev.org/zuul/zuul-client

Change-Id: I0da91b3e80ab9ccb134b9a37f88305fb24533b38