Don't clear branch cache
This keeps the branch cache continuously populated.  When part of
the cache is invalidated by a branch creation/deletion event, instead
of clearing the cache, this updates the data.

That can be slightly more efficient in that it may reduce the number
of ZK transactions, and it may also help us when adding layout loading
to zuul-web, in that those connections will not need to make any remote
calls to the code review system.

Change-Id: I0ad3ea0ecab50ae5261f9510e34ffb27ca83ac1e