Update patch set 1
Patch Set 1:

I don't know that this is strictly necessary.

I'm pretty sure we're going to need to have zuul-web instantiate all of the connections in order to have a branch cache.  Given that, we could probably accept that they would make external calls to the code review system to resolve branches.

However, we may want to limit what zuul-web does as much as possible, and so even though we need to instantiate the connections, we may not want to expect (or allow) them to make remote calls.  If that's the case, then I do think we need this so that we can rely exclusively on data in ZK.

Patch-set: 1