Split up registerScheduler() and onLoad() methods
This is an early preparation step for removing the RPC calls between
zuul-web and the scheduler.

In order to do so we must initialize the ConfigLoader in zuul-web which
requires all connections to be available. Therefore, this change ensures
that we can load all connections in zuul-web without providing a
scheduler instance.

To avoid unnecessary traffic from a zuul-web instance the onLoad()
method now initializes connection event queues and event watchers only
if a scheduler instance is available on the connection.

Change-Id: I3c1d2995e81e17763ae3454076ab2f5ce87ab1fc