Update patch set 4
Patch Set 4:

I also think we weren't handling the whitelabel case right; in that case, we don't want to do an extra call to the tenant info endpoint because the main info endpoint already is the tenant info endpoint, so it has our auth information.  I've fixed that in the next PS too.

Finally, I kept the callback page as a functional component, however, our version of react-router-dom was too old to support hooks, so I bumped that in the next PS.

Patch-set: 4