Update patch set 5
Patch Set 5:

The two mouseovers of the pipelines types (one for the name, the other for the description) is a little confusing and seems a little redundant (for one thing, I keep mousing over the wrong one to see the description).  How about we remove the (?) icon, and make the mouseover for the pipeline type icon include the description?  So if you mouseover the flask it says "Independent pipeline: every change is independent of other changes..." or something similar.

Patch-set: 5
CC: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>